backend/requirements.txt
# Core & Framework
Django==4.2
djangorestframework==3.14.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.10

# Environment Management
python-dotenv==1.0.0
django-environ==0.12.0

# API & Documentation
django-cors-headers==4.7.0
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.4.1

# Asynchronous Tasks (Celery)
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
gevent==23.9.1

# Caching
django-redis==5.3.0

# Testing
pytest==8.2.0
pytest-django==4.8.0
factory-boy==3.3.0
freezegun==1.4.0

# Deployment & Performance
gunicorn==21.2.0
whitenoise==6.6.0
